+++ src/main/scala/krs/domain/DomainModule.scala
@@ -2,4 +2,5 @@ package krs.domain
 
 trait DomainModule {
   def userRepository: UserRepository
+  def partnerRepository: PartnerRepository
 }
