+++ src/main/scala/krs/infrastructure/InfrastructureModule.scala
@@ -6,5 +6,6 @@ import krs.api.{ ApiModule }
 trait InfrastructureModule { this: ApiModule with DomainModule =>
 
   override val userRepository: UserRepository = new UserRepositoryFS("./fixtures/users.json")
+  override val partnerRepository: PartnerRepository = new PartnerRepositoryFS("./fixtures/offers.json")
 
 }
